chore: improve base32.js
When I say 'improve', I mean it uses less lines of code now. It's not, strictly speaking, a better implementation. It just so happens that Javascript numbers can have a radix inserted for their toString() function call, and padding that is pretty easy too, so here we are: 11 lines to decode, 11 lines to encode.