first cmd testing
I'll work my way through the rest of the commands directory this way. A couple other things (import.js, export.js, printer.js\?) should also get tested, but a lot of it is kinda noncritical. The system is currently passing the test vectors from the relevant RFCs, which is the biggest deal.