remove tests
Pretty sure these tests aren't really accomplishing anything. The test vector tests for RFC 6238 and RFC 4648 are good, and some other miscellaneous testing might be helpful, but short of end-to-end testing I don't think unit tests for each and every .js file are going to be truly helpful for this software. Open to suggestions and counter-arguments.